refactored and added all pickers to main search screen
