--- client_app/lib/views/search/text_search_data.dart
@@ -1,56 +1,0 @@
-import 'package:flutter/material.dart';
-
-final textShapes = [
-  '원형',
-  '타원형',
-  '반원형',
-  '삼각형',
-  '사각형',
-  '마름모',
-  '장방형',
-  '오각형',
-  '육각형',
-  '팔각형',
-  '기타'
-];
-
-final textColorNames = [
-  '햐양',
-  '노랑',
-  '주황',
-  '분홍',
-  '빨강',
-  '갈색',
-  '연두',
-  '초록',
-  '청록',
-  '파랑',
-  '남색',
-  '자주',
-  '보라',
-  '회색',
-  '검정',
-  '투명'
-];
-
-const textColorValues = <Color>[
-  Color(0xFFFFFFFF),
-  Color(0xFFE9DA49),
-  Color(0xFFE1642A),
-  Color(0xFFDF5382),
-  Color(0xFFC6281C),
-  Color(0xFF741811),
-  Color(0xFF86C93D),
-  Color(0xFF3D8A33),
-  Color(0xFF55ACC3),
-  Color(0xFF3B8FD3),
-  Color(0xFF06256D),
-  Color(0xFFAD2A95),
-  Color(0xFF6E1EA1),
-  Color(0xFF7B7B7B),
-  Color(0xFF000000),
-  Color(0xFFC8D5E7),
-];
-
-final textCharacters = ['정제', '경질', '연질'];
-final textDividers = ['+', '-', '기타'];
